+++ public/locales/en/translation.json
@@ -81,6 +81,10 @@
     "view": "View",
     "viewProducts": "View products",
     "requiredField": "Reqiered field",
+    "onlyLetters": "Include only letters",
+    "incorrectNumber": "Your phone number is incorrect",
+    "emailPattern": "Only english, digits and special characters allowed",
+    "incorrectEmail": "Please enter a valid email",
     "emptyBasket": "Basket is Empty",
     "test": "test",
     "test1": "test1",
