+++ public/locales/fr/translation.json
@@ -78,5 +78,10 @@
     "scheduleWeekend": "Dimanche : Fermé",
     "shippingAndPayment": "Livraison et paiement",
     "emptyBasket": "Le panier est vide",
-    "returnBtn": "Return"
+    "returnBtn": "Return",
+    "requiredField": "Reqiered field",
+    "onlyLetters": "Include only letters",
+    "incorrectNumber": "Your phone number is incorrect",
+    "emailPattern": "Only english, digits and special characters allowed",
+    "incorrectEmail": "Please enter a valid email"
 }
