+++ public/locales/ru/translation.json
@@ -79,6 +79,10 @@
     "view": "Смотреть",
     "viewProducts": "Смотреть товары",
     "requiredField": "Обязательное поле",
+    "onlyLetters": "Только буквы",
+    "incorrectNumber": "Номер телефона включает в себя только цифры",
+    "emailPattern": "Только латинские буквы, цифры и специальные символы",
+    "incorrectEmail": "Введите правильный Email",
     "test1": "test1",
     "test2": "test2",
     "test3": "test3",
